+++ FirebaseAppCheck/Sources/Public/FirebaseAppCheck/FirebaseAppCheck.h
@@ -30,3 +30,7 @@
 
 // App Attest provider.
 #import "FIRAppAttestProvider.h"
+
+// Recaptcha Enterprise provider
+#import "FIRAppCheckRecaptchaEnterpriseProviderFactory.h"
+#import "FIRRecaptchaEnterpriseProvider.h"