+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [added] Added reCAPTCHA Enterprise provider.
+
 # 10.27.0
 - [fixed] [CocoaPods] missing symbol error for FIRGetLoggerLevel. (#12899)